Support pip_lock_down in repo_server role
The repo_server role does not use pip_lock_down which allows this role,
repo_build, to install packages before the repo is built or updated. The
impact of this is that any additional roles e.g. logging or monitoring
that a deployer may wish to apply to the repo servers are not restricted
to the same package list as all other hosts.

This commit adds the isolated flag to the pip commands required to run
before the repo exists. This causes those commands to ignore the
pip.conf file that is created by pip_lock_down.

In addition, the role galera_client has the var
galera_client_pip_packages set to an empty list to prevent it installing
and pip packages. The galera_client role is a dependency because it is
used to ensure the appropriate MariaDB apt libraries are installed for
use when building the MySQL-Python wheel.

An additional commit will be introduced into repo_build to add
pip_lock_down and will depend on this commit.

Change-Id: I2fbcb76f349242bf45161a1789891ed3133557eb